requirements.txt
# Requirements for Hugging Face Spaces deployment
# Python 3.11 | Streamlit SDK
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
imbalanced-learn==0.11.0
tensorflow==2.15.1
opencv-python-headless==4.8.1.78
networkx==3.2.1
shap==0.44.1
matplotlib==3.8.2
streamlit==1.29.0
python-dotenv==1.0.0
joblib==1.3.2
